fix: move menu reorg logic from crud app into Menu component
